protect uesr routes

+++ resources/users/routes.py
@@ -1,4 +1,5 @@
 from flask.views import MethodView
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from flask_smorest import abort
 from sqlalchemy.exc import IntegrityError
 
@@ -14,25 +15,23 @@ class UserList(MethodView):
     users = UserModel.query.all()
     return users
 
+  @jwt_required()
   @bp.arguments(AuthUserSchema)
   def delete(self, user_data):
-    user = UserModel.query.filter_by(username=user_data['username']).first()
-    if user and user.check_password(user_data['password']):
+    user_id = get_jwt_identity()
+    user = UserModel.query.get(user_id)
+    
+    if user and user.username == user_data['username'] and user.check_password(user_data['password']):
       user.delete()
       return {'message':f'{user_data["username"]} deleted'}, 202
     abort(400, message='Username or Password Invalid')
 
-@bp.route('/user/<user_id>')
-class User(MethodView):
-
-  @bp.response(200, UserSchemaNested)
-  def get(self, user_id):
-    user = UserModel.query.get_or_404(user_id, description='User Not Found')
-    return user
-
+    
+  @jwt_required()
   @bp.arguments(UpdateUserSchema)
   @bp.response(202, UserSchema)
-  def put(self, user_data, user_id):
+  def put(self, user_data):
+    user_id = get_jwt_identity()
     user = UserModel.query.get_or_404(user_id, description='User Not Found')
     if user and user.check_password(user_data['password']):
       try:
@@ -42,12 +41,22 @@ class User(MethodView):
       except IntegrityError:
         abort(400, message='Username or Email already Taken')
 
+@bp.route('/user/<user_id>')
+class User(MethodView):
 
-@bp.route('/user/follow/<follower_id>/<followed_id>')
+  @bp.response(200, UserSchemaNested)
+  def get(self, user_id):
+    user = UserModel.query.get_or_404(user_id, description='User Not Found')
+    return user
+
+
+@bp.route('/user/follow/<followed_id>')
 class FollowUser(MethodView):
   
+  @jwt_required()
   @bp.response(200, UserSchema(many=True))
-  def post(self, follower_id, followed_id):
+  def post(self, followed_id):
+    follower_id = get_jwt_identity()
     user = UserModel.query.get(follower_id)
     user_to_follow = UserModel.query.get(followed_id)
     if user and user_to_follow:
@@ -55,7 +64,9 @@ class FollowUser(MethodView):
       return user.followed.all()
     abort(400, message='Invalid user info')
 
-  def put(self, follower_id, followed_id):
+  @jwt_required()
+  def put(self, followed_id):
+    follower_id = get_jwt_identity()
     user = UserModel.query.get(follower_id)
     user_to_unfollow = UserModel.query.get(followed_id)
     if user and user_to_unfollow:
